Monocle WS Driver render loop functional.
